+++ src/network_time.rs
@@ -185,7 +185,7 @@ impl NtpTime {
         let (year, doy, hour, minute, second) = unix_seconds_to_ymd(unix_secs);
 
         let mut abs = AbsoluteTime::new(doy, hour, minute, second, nanos)?;
-        abs.year = Some(year);
+        abs.set_year(Some(year));
         Ok(abs)
     }
 
@@ -278,7 +278,7 @@ impl PtpTime {
         let (year, doy, hour, minute, second) = unix_seconds_to_ymd(utc_secs);
 
         let mut abs = AbsoluteTime::new(doy, hour, minute, second, self.nanoseconds)?;
-        abs.year = Some(year);
+        abs.set_year(Some(year));
         Ok(abs)
     }